Breaking out the profile page more
