+++ routes/userRoutes.js
@@ -110,8 +110,9 @@ module.exports = app => {
         res.status(200).send(`success`)
     });
 
-    app.get(`/api/getAvatar/:id`, async (req, res) => {
+    app.get(`/api/avatar/:id`, async (req, res) => {
         const { id } = req.params;
+        console.log(id)
         const avatar = await s3Handler.getAvatar(id);
         res.status(200).send(avatar);
     });
